+++ module-system/sov-modules-macros/tests/constants/bech32_constant_not_a_string.rs
@@ -1,10 +1,7 @@
 use sov_bank::TokenId;
-use sov_modules_api::macros::config_bech32_constant;
+use sov_modules_api::macros::config_bech32;
 
-
-#[config_bech32_constant]
-const TEST_ARRAY_OF_U8: TokenId;
-
+const TEST_ARRAY_OF_U8: TokenId = config_bech32!("TEST_ARRAY_OF_U8", TokenId);
 
 fn main() {
     assert_eq!(TEST_ARRAY_OF_U8, TokenId::from([0; 32]));