+++ module-system/sov-modules-macros/tests/constants/bech32_constant_prefix_too_long.rs
@@ -1,12 +1,9 @@
-use sov_modules_api::macros::config_bech32_constant;
 use sov_modules_api::impl_hash32_type;
-
+use sov_modules_api::macros::config_bech32;
 
 impl_hash32_type!(MyTokenId, MyTokenBech, "tok");
 
-#[config_bech32_constant]
-const TEST_TOKEN_ID: MyTokenId;
-
+const TEST_TOKEN_ID: MyTokenId = config_bech32!("TEST_TOKEN_ID", MyTokenId);
 
 fn main() {
     assert_eq!(TEST_TOKEN_ID, MyTokenId::from([0; 32]));